feat: Supabase 실데이터 환경 완전 연동 및 주문 시스템 구축
## 주요 변경사항

### 🗄️ Supabase PostgreSQL 연동
- Flyway 마이그레이션 V100~V113 순서 정리
- orders/order_items 테이블 최소 구조로 안정화
- SMTP 설정 수정 (eyub ylej wyuq nqvj)
- AdminInitializationRunner supabase 프로파일 제외

### 🛡️ 보안 및 인증 강화
- CORS 설정 보안 강화 (allowCredentials=true 호환)
- JWT 토큰 생성/검증 정상 동작 확인
- 사용자 타입별 권한 체계 구축

### 📋 주문 관리 시스템
- SimpleOrderController 구현 (MVP 버전)
- OrderItem 엔티티 간소화 (NPE 방지)
- 비즈니스 규칙 예외 처리 추가
- 상태 전환 검증 로직 구현

### 🔧 개발 환경 개선
- 통합 테스트 환경 구축
- 프론트엔드 주문 폼 컴포넌트 추가
- E2E 테스트 스크립트 구현

## 테스트 결과
- ✅ 사용자 생성/인증 성공 (3개 사용자 생성)
- ✅ 백엔드 서버 정상 실행 (포트 8080)
- ✅ 프론트엔드 서버 정상 실행 (포트 3008)
- ✅ 실데이터 환경에서 E2E 플로우 가능

+++ JJYCS/backend/src/main/java/com/ysc/lms/service/OrderBusinessRuleService.java
@@ -126,7 +126,7 @@ public class OrderBusinessRuleService {
                 BigDecimal itemCbm = item.getWidth()
                     .multiply(item.getHeight())
                     .multiply(item.getDepth())
-                    .multiply(item.getQuantity())
+                    .multiply(BigDecimal.valueOf(item.getQuantity()))
                     .divide(BigDecimal.valueOf(1000000), 6, RoundingMode.HALF_UP);
                 
                 totalCbm = totalCbm.add(itemCbm);
@@ -150,7 +150,7 @@ public class OrderBusinessRuleService {
         if (order.getOrderItems() != null) {
             for (OrderItem item : order.getOrderItems()) {
                 if (item.getUnitPrice() != null && item.getQuantity() != null) {
-                    BigDecimal itemTotalValue = item.getUnitPrice().multiply(item.getQuantity());
+                    BigDecimal itemTotalValue = item.getUnitPrice().multiply(BigDecimal.valueOf(item.getQuantity()));
                     totalThbValue = totalThbValue.add(itemTotalValue);
                     log.debug("Item THB value: {} x {} = {}", item.getUnitPrice(), item.getQuantity(), itemTotalValue);
                 }
